# src/main/resources/static/api-spec.yml
openapi: 3.0.0
info:
  title: FiSion Apps
  version: 1.0.0
servers:
  - url: http://localhost:8080
  - url: http://{{base_url}}
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...
paths:
  /api/getEncryptedPassword:
    get:
      tags:
        - Login Page
      summary: getEncryptedPassword
      parameters:
        - name: password
          in: query
          schema:
            type: string
          example: password
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
  /api/sendResetPassword:
    post:
      tags:
        - Login Page
      summary: Reset Password
      requestBody:
        content: {}
      security:
        - bearerAuth: []
      parameters:
        - name: email
          in: query
          schema:
            type: string
          example: lorddev.comp@gmail.com
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
  /api/login:
    post:
      tags:
        - Login Page
      summary: Login
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                username: username
                password: password
      parameters:
        - name: ''
          in: header
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
  /api/authenticate:
    post:
      tags:
        - Login Page
      summary: Authenticate JWT
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                username: username
                password: password
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
...